Overworld objects can now execute on-sight scripts, persist across map loads, and start battles with the player.

+++ Assets/Scripts/src/XP.cs
@@ -7,35 +7,35 @@ public static class XP
     {
         return level switch
         {
-            < 50 => (int)(level * level * level * (100 - level) / 50),
-            < 68 => (int)(level * level * level * (150 - level) / 100),
-            < 98 => (int)(level * level * level * ((1911 - 10 * level) / 3) / 500),
-            _ => (int)(level * level * level * (160 - level) / 100),
+            < 50 => (level * level * level * (100 - level) / 50),
+            < 68 => (level * level * level * (150 - level) / 100),
+            < 98 => (level * level * level * ((1911 - 10 * level) / 3) / 500),
+            _ => (level * level * level * (160 - level) / 100),
         };
     }
     public static int FastXP(int level)
     {
-        return (int)(4 * level * level * level / 5);
+        return (4 * level * level * level / 5);
     }
     public static int MediumFastXP(int level)
     {
-        return (int)(level * level * level);
+        return (level * level * level);
     }
     public static int MediumSlowXP(int level)
     {
-        return (int)(Max(0, (level * (100 + (level * ((6 * level / 5) - 15)))) - 140));
+        return (Max(0, (level * (100 + (level * ((6 * level / 5) - 15)))) - 140));
     }
     public static int SlowXP(int level)
     {
-        return (int)(5 * level * level * level / 4);
+        return (5 * level * level * level / 4);
     }
     public static int FluctuatingXP(int level)
     {
         return level switch
         {
-            >= 36 => (int)(level * level * level * ((level / 2) + 32) / 50),
-            >= 15 => (int)(level * level * level * (level + 14) / 50),
-            _ => (int)(level * level * level * (((level + 1) / 3) + 24) / 50)
+            >= 36 => (level * level * level * ((level / 2) + 32) / 50),
+            >= 15 => (level * level * level * (level + 14) / 50),
+            _ => (level * level * level * (((level + 1) / 3) + 24) / 50)
         };
     }
     public static int LevelToXP(int level, XPClass XPClass)
